[TWINFACES-413] deleted transition-alias Api
