--- src/entities/transition-alias/api/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from "./use-search";
