--- src/entities/transition-alias/libs/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from "./use-select-adapter";
