--- src/entities/transition-alias/libs/index.ts
@@ -1,1 +1,0 @@
-export * from "./hooks";
