+++ src/entities/twin-flow-transition/api/hooks/index.ts
@@ -4,3 +4,4 @@ export * from "./use-search-triggers";
 export * from "./use-search-validator-rules";
 export * from "./use-create-transition";
 export * from "./use-update-transition";
+export * from "./use-search-alias";
