+++ src/entities/twin-flow-transition/libs/constants.ts
@@ -1,15 +1,16 @@
 import { z } from "zod";
-import { FIRST_ID_EXTRACTOR } from "@/shared/libs";
+
+import {
+  FIRST_ID_EXTRACTOR,
+  FIRST_TWIN_FLOW_ID_EXTRACTOR,
+} from "@/shared/libs";
 
 import { FIRST_ID_EXTRACTOR } from "@/shared/libs";
 
 export const ENTITY_COLOR = "#8B5CF6"; // text-violet-500
 
 export const TWIN_FLOW_TRANSITION_SCHEMA = z.object({
-  twinflow: z
-    .array(z.object({ id: z.string().uuid(), twinClassId: z.string() }))
-    .min(1, "Twinflow can not be empty")
-    .nullable(),
+  twinflow: FIRST_TWIN_FLOW_ID_EXTRACTOR,
   alias: z
     .array(z.object({ alias: z.string() }))
     .min(1, "Alias can not be empty")
