+++ src/entities/twin-flow-transition/libs/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from "./use-filters";
+export * from "./use-alias-select-adapter";
