+++ src/features/api/context/private-context-provider.tsx
@@ -41,6 +41,15 @@ import { createFeaturerApi, FeaturerApi } from "@/entities/featurer";
 import { createLinkApi, LinkApi } from "@/entities/link";
 import { createPermissionApi, PermissionApi } from "@/entities/permission";
 import {
+  PipelineStepApi,
+  createPipelineStepApi,
+} from "@/entities/factory-pipeline-step";
+import { FeaturerApi, createFeaturerApi } from "@/entities/featurer";
+import { LinkApi, createLinkApi } from "@/entities/link";
+import { PermissionApi, createPermissionApi } from "@/entities/permission";
+import {
+  PermissionGroupApi,
+  createPermissionGroupApi,
   PermissionGroupApi,
   createPermissionGroupApi,
 } from "@/entities/permission-group";
@@ -64,7 +73,9 @@ import {
   TwinFlowTransitionApi,
   createTwinFlowTransitionApi,
 } from "@/entities/twin-flow-transition";
+import { TwinStatusApi, createTwinStatusApi } from "@/entities/twin-status";
 import {
+  TwinFlowSchemaApi,
   PermissionTwinRoleApi,
   createPermissionTwinRoleApi,
 } from "@/entities/twin-role";
@@ -73,26 +84,19 @@ import {
   TwinFlowSchemaApi,
   createTwinFlowSchemaApi,
 } from "@/entities/twinFlowSchema";
+import {
+  PermissionTwinRoleApi,
+  createPermissionTwinRoleApi,
+} from "@/entities/twinRole";
+import { UserApi, createUserApi } from "@/entities/user";
+import { UserGroupApi, createUserGroupApi } from "@/entities/userGroup";
+import { ApiContext, ApiSettings } from "@/shared/api";
 import { UserApi, createUserApi } from "@/entities/user";
 import { UserGroupApi, createUserGroupApi } from "@/entities/user-group";
 import { ApiSettings, PrivateApiContext, TwinsAPI } from "@/shared/api";
 import { LoadingOverlay } from "@/shared/ui";
-import { env } from "next-runtime-env";
-import createClient from "openapi-fetch";
-import React from "react";
+
 import { useAuthUser } from "../auth";
-import {
-  createFactoryMultiplierApi,
-  FactoryMultiplierApi,
-} from "@/entities/factory-multiplier";
-import {
-  createTransitionAliasApi,
-  TransitionAliasApi,
-} from "@/entities/transition-alias";
-import {
-  createPipelineStepApi,
-  PipelineStepApi,
-} from "@/entities/factory-pipeline-step";
 
 import { useAuthUser } from "../../auth";
